Move hidden class code into a separate file.
Since it's one class and it's big enough, might as well have it's own
file.
